requirements.txt
absl-py==1.4.0
astor==0.8.1
attrs==22.2.0
blis==0.2.4
cached-property==1.5.2
certifi==2023.5.7
charset-normalizer==2.0.12
click==8.0.4
cymem==2.0.7
dataclasses==0.8
gast==0.5.4
grpcio==1.48.2
h5py==2.10.0
idna==3.4
importlib-metadata==4.8.3
importlib-resources==5.4.0
joblib==1.1.1
jsonschema==3.0.2
Keras==2.2.0
Keras-Applications==1.0.2
keras-layer-normalization==0.13.0
Keras-Preprocessing==1.0.1
Markdown==3.3.7
murmurhash==1.0.9
nltk==3.6.7
numpy==1.19.5
plac==0.9.6
preshed==2.0.1
protobuf==3.19.6
pyrsistent==0.18.0
PyYAML==6.0
regex==2023.5.5
requests==2.27.1
scipy==1.5.4
six==1.16.0
spacy==2.1.4
srsly==1.0.6
tensorboard==1.10.0
tensorflow==1.10.0
termcolor==1.1.0
thinc==7.0.8
tqdm==4.64.1
typing-extensions==4.1.1
urllib3==1.26.16
wasabi==0.10.1
Werkzeug==2.0.3
zipp==3.6.0
